pyproject.toml
[project]
name = "screen-recorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nats-py>=2.14.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "typer>=0.24.1",
]

[project.scripts]
screen-record = "screen_recorder.__main__:app"

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"